Rename GradleSettingsWriting.kt and rename a function
